Link files in `dart_library` instead of `layout_action`
The only files we will be allowed to import for packages are those under
'lib/'. Some of these files may be generated so we create a _merged_lib/
directory to replace 'lib/' which is split between the source tree and
generated stuff in bazel-bin.

We must create the _merged_lib for all dart libraries, regardless of
whether they have no generated sources, because they may get merged with
a target that does have generated sources.

For files that are not under 'lib/' we don't need to create a link since
they will not be imported via package: uri.

Pass in a package-spec path rather than the File to the
package_spec_action. the relative paths inside .packages are relative
*to that file* so when the path to the package_spec is nested in a
directory the number "../" we need can vary - we need to know the path
to the file not just the path to the target creating the file

Inside dart_library always create a _merged_lib directory and link in
files from lib/. This link will point to something either under
bazel-bin or the source tree, but later actions that need to read a
package: url won't care. This also centralizes what used to happen
multiple times as differect targets needed a layout_action covering the
same library.